+++ NEWS.md
@@ -1,8 +1,10 @@
-# spAbundance 0.3.0
+# spAbundance 0.2.2
 
 + Added in the `loadings.diag` argument to `simMsAbund()` to allow for simulating data where the diagonal of the factor loadings matrix is set to 1 (i.e., the default and what was done in previous versions) or to random positive values. 
 + Updated `ppcAbund()` to work with `svcAbund()` models (Poisson and NB).
-+ Fixed a bug in all GLMM models that prevented the models from running when a site-level offset was supplied into the model. Thanks to Sonia Illanas for brining this to my attention.
++ Fixed a bug in all GLMM models that prevented the models from running when a site-level offset was supplied into the model. Thanks to Sonia Illanas for bringing this to my attention.
++ Updated `simAbund()` to allow simulation of Gaussian and ZI-Gaussian data with multiple replicate samples per site. 
++ Updated `svcAbund()` to allow fitting Gaussian SVC models with multiple replicate samples per site. 
 
 # spAbundance 0.2.1